Removed hardcoded path relative root
The generated paths always had a hardcoded `./` prefix. This resulted in
paths like this:
`import IWow from './../foo/bar/IWow';`
While technically correct, the initial `./` is redundant, and looks a
bit silly.
I've removed the hardcoded `./` and added a check for the first
character of the relative path being a `.`. This means it will not be
added, unless the import is for a file in the same directory as the
source file, in which case it will result in `./IWow`